+++ src/Query.hs
@@ -98,6 +98,6 @@ execute args state events query = do
       putStrLn "unfog <help|h>"
       putStrLn "unfog <version|v>"
 
-    ShowVersion                 -> printVersion rtype "0.2.0"
+    ShowVersion                 -> printVersion rtype "0.2.1"
 
     Query.Error command message -> printErr rtype $ command ++ ": " ++ message