+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2019-12-08
+
 ### Added
 
 - Add basic help command [#3]
@@ -73,7 +75,8 @@ First release :tada:
 
 - Linux binaries
 
-[unreleased]: https://github.com/unfog-io/unfog-cli/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/unfog-io/unfog-cli/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/unfog-io/unfog-cli/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/unfog-io/unfog-cli/compare/v0.1.4...v0.2.0
 [0.1.4]: https://github.com/unfog-io/unfog-cli/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/unfog-io/unfog-cli/compare/v0.1.2...v0.1.3